Added WeatherFragment backend connection to web service.
