+++ app/src/main/java/edu/uw/chather/MainActivity.java
@@ -1,0 +1,66 @@
+package edu.uw.chather;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
+import androidx.navigation.ui.AppBarConfiguration;
+import androidx.navigation.ui.NavigationUI;
+
+import android.os.Bundle;
+
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+import edu.uw.chather.MainActivityArgs;
+import edu.uw.chather.ui.model.UserInfoViewModel;
+
+
+public class MainActivity extends AppCompatActivity {
+
+    private AppBarConfiguration mAppBarConfiguration;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        MainActivityArgs args = MainActivityArgs.fromBundle(getIntent().getExtras());
+
+        //String email = args.getJwt();
+        //Import com.auth0.android.jwt.JWT
+        //JWT jwt = new JWT(args.getJwt());
+
+        // Check to see if the web token is still valid or not. To make a JWT expire after a
+        // longer or shorter time period, change the expiration time when the JWT is
+        // created on the web service.
+       // if(!jwt.isExpired(0)) {
+        //    new ViewModelProvider(this, new UserInfoViewModel.UserInfoViewModelFactory(jwt)).get(UserInfoViewModel.class);
+        //} else {
+            //In production code, add in your own error handling/flow for when the JWT is expired
+        //    throw new IllegalStateException("JWT is expired!");
+        //}
+
+        new ViewModelProvider(this,
+                new UserInfoViewModel.UserInfoViewModelFactory(args.getEmail(), args.getJwt())
+                    ).get(UserInfoViewModel.class);
+
+        BottomNavigationView navView = findViewById(R.id.nav_view);
+        // Passing each menu ID as a set of Ids because each
+        // menu should be considered as top level destinations.
+        mAppBarConfiguration = new AppBarConfiguration.Builder(
+                R.id.success, R.id.navigation_chats, R.id.navigation_connections,  R.id.weatherFragment)
+                .build();
+        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
+        NavigationUI.setupActionBarWithNavController(this, navController, mAppBarConfiguration);
+        NavigationUI.setupWithNavController(navView, navController);
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
+        return NavigationUI.navigateUp(navController, mAppBarConfiguration)
+                || super.onSupportNavigateUp();
+    }
+
+
+}
