+++ app/src/main/java/edu/uw/chather/Retrofit/RetrofitClient.java
@@ -1,0 +1,19 @@
+package edu.uw.chather.Retrofit;
+
+import retrofit2.Retrofit;
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class RetrofitClient {
+    private static Retrofit instance;
+
+    public static Retrofit getInstance() {
+        if(instance == null) {
+            instance = new Retrofit.Builder()
+                    .baseUrl("https://api.openweathermap.org/data/2.5/")
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
+                    .build();
+        }
+    }
+}
