--- app/src/main/java/edu/uw/tcss450/AuthActivity.java
@@ -1,56 +1,0 @@
-package edu.uw.tcss450;
-/**
- * Duy Nguyen
- * TCSS 450
- * Lab 1 Assignment
- */
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-import android.util.Log;
-
-/**
- * This is where to start all the main activities.
- */
-public class AuthActivity extends AppCompatActivity {
-
-    private static final String TAG = "MyActivity";
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_auth);
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        Log.i(TAG, "onStart");
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        Log.d(TAG,"onResume");
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        Log.w(TAG, "onPause");
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        Log.v(TAG, "onStop");
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        Log.e(TAG, "onDestroy");
-
-    }
-}
