--- app/src/main/java/edu/uw/tcss450/ui/model/UserInfoViewModel.java
@@ -1,49 +1,0 @@
-package edu.uw.tcss450.ui.model;
-
-import androidx.annotation.NonNull;
-import androidx.lifecycle.ViewModel;
-import androidx.lifecycle.ViewModelProvider;
-
-import com.auth0.android.jwt.JWT;
-
-public class UserInfoViewModel extends ViewModel {
-
-    private final String mEmail;
-    private static String mJwt;
-
-    private UserInfoViewModel(String email, String jwt) {
-        mEmail = email;
-        mJwt = jwt;
-    }
-
-    public String getEmail() {
-        return mEmail;
-    }
-
-    public static String getmJwt() {
-        return mJwt;
-    }
-
-    public static class UserInfoViewModelFactory implements ViewModelProvider.Factory {
-
-        private final String email;
-        private final String jwt;
-
-        public UserInfoViewModelFactory(String email, String jwt) {
-            this.email = email;
-            this.jwt = jwt;
-        }
-
-        @NonNull
-        @Override
-        public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-            if (modelClass == UserInfoViewModel.class) {
-                return (T) new UserInfoViewModel(email, jwt);
-            }
-            throw new IllegalArgumentException(
-                    "Argument must be: " + UserInfoViewModel.class);
-        }
-    }
-}
-
-
