--- app/src/main/java/edu/uw/tcss450/ui/weather/WeatherFragment.java
@@ -1,22 +1,0 @@
-package edu.uw.tcss450.ui.weather;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import edu.uw.tcss450.R;
-
-public class WeatherFragment extends Fragment {
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_weather, container, false);
-    }
-}
