+++ app/src/main/res/navigation/nav_graph.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/signInFragment"
-        android:name="edu.uw.tcss450.ui.signin.SignInFragment"
+        android:name="edu.uw.chather.ui.signin.SignInFragment"
         android:label="fragment_sign_in"
         tools:layout="@layout/fragment_sign_in" >
         <action
@@ -28,7 +28,7 @@
     </fragment>
     <fragment
         android:id="@+id/registerFragment"
-        android:name="edu.uw.tcss450.ui.register.RegisterFragment"
+        android:name="edu.uw.chather.ui.register.RegisterFragment"
         android:label="fragment_register"
         tools:layout="@layout/fragment_register" >
         <action
@@ -37,7 +37,7 @@
     </fragment>
     <activity
         android:id="@+id/mainActivity"
-        android:name="edu.uw.tcss450.MainActivity"
+        android:name="edu.uw.chather.MainActivity"
         android:label="MainActivity" >
         <argument
             android:name="email"
